Add water supply form

+++ lib/widgets/water-daily-stats.dart
@@ -9,7 +9,7 @@ class WaterDailyStats extends StatelessWidget {
     @required this.target,
   }) : super(key: key);
 
-  final double current;
+  final int current;
   final int target;
 
   @override
